backend-app/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
databases
httpx
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
psycopg2-binary


